// src/types/projectEntry.ts
/*-----------------------------------------------------------------------------------------------
 *  Copyright (c) Zulfazli (fazelstudio). All rights reserved.
 *  Licensed under the MIT License. See LICENSE file in the project root for license information.
 *-----------------------------------------------------------------------------------------------*/

export interface ProjectEntry {
  id: string;
  name: string;
  folderPath: string;
  createdAt: string;
  lastOpenedAt: string;
}